backend/requirements/base.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
psycopg[binary]==3.3.3
pgvector==0.3.5
redis==5.0.8
kombu==5.3.4
Pillow==10.4.0
drf-spectacular==0.27.2
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.7.0
django-debug-toolbar==4.4.2
requests==2.32.3
